source/backend/app/requirements.txt
aniso8601==9.0.1
appdirs==1.4.4
audioread==3.0.0
azure-ai-textanalytics==5.2.0b4
azure-cognitiveservices-speech==1.24.0
azure-common==1.1.28
azure-core==1.26.0
black==22.10.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
decorator==5.1.1
Flask==2.2.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.2
greenlet==1.1.3.post0
idna==3.4
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
librosa==0.9.2
llvmlite==0.39.1
MarkupSafe==2.1.1
msrest==0.7.1
mypy-extensions==0.4.3
numba==0.56.3
numpy==1.23.4
oauthlib==3.2.2
packaging==21.3
pathspec==0.10.1
platformdirs==2.5.2
pooch==1.6.0
pycparser==2.21
pydub==0.25.1
pyparsing==3.0.9
pytz==2022.5
requests==2.28.1
requests-oauthlib==1.3.1
resampy==0.4.2
scikit-learn==1.1.3
scipy==1.9.3
six==1.16.0
soundfile==0.11.0
SQLAlchemy==1.4.42
threadpoolctl==3.1.0
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.12
webrtcvad==2.0.10
Werkzeug==2.2.2
